api/requirements.txt
fastapi
uvicorn
python-jose
passlib
pydantic
pydantic-settings
httpx
